+++ dom/prototype/PrototypeDocumentContentSink.cpp
@@ -415,6 +415,17 @@ void PrototypeDocumentContentSink::CloseElement(Element* aElement) {
 }
 
 nsresult PrototypeDocumentContentSink::ResumeWalk() {
+  nsresult rv = ResumeWalkInternal();
+  if (NS_FAILED(rv)) {
+    nsContentUtils::ReportToConsoleNonLocalized(
+        NS_LITERAL_STRING("Failed to load document from prototype document."),
+        nsIScriptError::errorFlag, NS_LITERAL_CSTRING("Prototype Document"),
+        mDocument, mDocumentURI);
+  }
+  return rv;
+}
+
+nsresult PrototypeDocumentContentSink::ResumeWalkInternal() {
   MOZ_ASSERT(mStillWalking);
   // Walk the prototype and build the delegate content model. The
   // walk is performed in a top-down, left-to-right fashion. That