+++ layout/svg/nsSVGIntegrationUtils.h
@@ -69,6 +69,12 @@ class nsSVGIntegrationUtils final {
   static bool UsingMaskOrClipPathForFrame(const nsIFrame* aFrame);
 
   /**
+   * Returns true if the element has a clippath that is simple enough to
+   * be represented without a mask in WebRender.
+   */
+  static bool UsingSimpleClipPathForFrame(const nsIFrame* aFrame);
+
+  /**
    * Returns the size of the union of the border-box rects of all of
    * aNonSVGFrame's continuations.
    */